requirements.txt
aiogram~=3.2.0
python-dotenv~=1.0.0
SQLAlchemy~=2.0.23
alembic~=1.13.0
psycopg2-binary~=2.9.9
asyncpg~=0.29.0
